Fix #1: When deleting a single credential, verify with user before deleting.

+++ ward/del.go
@@ -28,6 +28,11 @@ func (app *App) runDel(query []string) {
     return
   }
 
-  db.DeleteCredential(credential)
-  app.printSuccess("Credential deleted.\n")
+  identifier := formatCredential(credential)
+  if confirm := app.readYesNo("Delete " + identifier); confirm {
+    db.DeleteCredential(credential)
+    app.printSuccess("Credential deleted.\n")
+  } else {
+    app.printError("Canceled.\n")
+  }
 }
